Added jwt on application.yml

+++ Application/machine-rental/src/main/java/rent/tycoon/persistance/config/security/token/AccessToken.java
@@ -5,7 +5,7 @@ import java.util.Set;
 public interface AccessToken {
     String getSubject();
 
-    Long getUserId();
+    int getUserId();
 
     Set<String> getRoles();
